% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowbcsgain.R
\name{bcs.gain.cow}
\alias{bcs.gain.cow}
\title{Nutrient Requirements to Increase Cow Body Condition Score}
\usage{
bcs.gain.cow(breed.code = 1, mbw.lbs = 1200, initial.bcs = 4,
  target.bcs = 6, first.calf = FALSE)
}
\arguments{
\item{breed.code}{Default is 1 = Angus.  Breed code must match a code listed
in the table 19_1 data file.}

\item{mbw.lbs}{Default is 1200.  This value represents the mature, unshrunk bw
(lbs) for cows > 3 yrs old.  For first calf heifers, calving at 2 yr is assumed.
mbw.lbs is the estimated mature weight for first calf heifers, and equations are based on
a shrunk bw of 0.85(mbw) at first calving and 0.9(mbw) at second calving.}

\item{initial.bcs}{Default is 4. This value represents the body condition score at
the end of the 2nd trimester or at calf weaning.}

\item{target.bcs}{Default is 6. This value represents the target body condition score
at calving.}

\item{first.calf}{Default is FALSE.  Change to TRUE to adjust requirements for first
calf heifers.}
}
\value{
Three tables are printed including nutrient requirement totals, nutrient
requirement as a proportion of dry matter intake, and a third table whereby crude
protein is adjusted to a ratio of 6:1 for tdn:cp.
}
\description{
\code{bcs.gain.cow} Nutrient requirements to increase body condition score of mature cows
or previously Calved first-calf heifers during the 3rd trimester following weaning
}
\examples{
bcs.gain.cow()
bcs.gain.cow(target.bcs=5)
bcs.gain.cow(first.calf=TRUE)

}
