% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingcattle.R
\name{growing.cattle}
\alias{growing.cattle}
\title{Nutrient Requirements for All Growing Cattle.}
\usage{
growing.cattle(breed.code = 1, Bull = FALSE, mbw.lbs = 1200,
  yearling = FALSE, target.adg = 2, from.bw.lbs = 400, to.bw.lbs = 800,
  increment.bw.lbs = 100, implant = TRUE, ionophore = TRUE,
  monensin = TRUE)
}
\arguments{
\item{breed.code}{Default is 1 = Angus.  Breed code must match a code listed
in the table 19_1 data file.}

\item{Bull}{Default is false. Adjustment for NEm applied when TRUE.}

\item{mbw.lbs}{Default is 1200.  This value represents the mature, unshrunk bw
(lbs) for growing cattle at finishing or at maturity.}

\item{yearling}{Default is FALSE. This value adjusts DMI for growing cattle > 11
months of age.}

\item{target.adg}{Default is 2. This value should represent realistic rates of
gain such as 0.5 as a low and 3.2 as a high.}

\item{from.bw.lbs}{Avoid using values < 400. This sets the minimum of the range
for prediction.}

\item{to.bw.lbs}{This value sets the maximum of the range for prediction.  Maximum
should be realistic for age (calf vs yearling).}

\item{increment.bw.lbs}{This value sets the incremental change between
\code{from.bw.lbs} to \code{to.bw.lbs}.}

\item{implant}{Default is TRUE.  When \code{Bull} is TRUE, implant is overridden to non-implant
factor of 0.94.}

\item{ionophore}{Default is TRUE. When \code{Bull} is TRUE, ionophore is overridden to
no ionophore.}

\item{monensin}{Default is TRUE.  When \code{ionophore} is FALSE, monensin is overridden.}
}
\value{
Two tables are printed including nutrient requirement totals and nutrient
requirement as a proportion of dry matter intake.
}
\description{
\code{growing.cattle} Nutrient requirements for growing cattle.
}
\examples{
growing.cattle()
growing.cattle(implant=FALSE,ionophore=FALSE)

}
