% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturecow.R
\name{mature.cow}
\alias{mature.cow}
\title{Calculate requirements for mature cows and first calf heifers}
\usage{
mature.cow(breed.code = 1, mbw.lbs = 1200, milk.pk.lb = 20, RMY = 5,
  first.calf = FALSE, first.calf.NEmadj = 1.25)
}
\arguments{
\item{breed.code}{Default is 1 = Angus.  Breed code must match a code listed
in the table 19_1 data file.}

\item{mbw.lbs}{Default is 1200.  This value represents the mature, unshrunk bw
(lbs) for cows > 3 yrs old.  For first calf heifers, calving at 2 yr is assumed.
mbw.lbs is the estimated mature weight for first calf heifers, and equations are based on
a shrunk bw of 0.85(mbw) at first calving and 0.9(mbw) at second calving.}

\item{milk.pk.lb}{Default is 20.  Users can modify peak milk (lbs). Peak milks is estimated
to occur at 8.5 weeks after calving.}

\item{RMY}{Default is 5.  Relative Milk Yield (RMY) is a convenient method of adjusting
milk production.}

\item{first.calf}{Default is FALSE.  Change to TRUE to adjust requirements for first
calf heifers.}

\item{first.calf.NEmadj}{Default is 1.25.  This value can be changed to NRC default 1.}
}
\value{
Two tables are printed including nutrient requirement totals and nutrient
requirement as a proportion of dry matter intake.
}
\description{
\code{mature.cow} returns calculated requirements for mature cows and first calf heifers as
daily totals and proportion of dry matter intake.
}
\examples{
mature.cow()
mature.cow(breed.code=4)
mature.cow(breed.code=4,first.calf=TRUE)
}
