% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bredheifer.R
\name{preg.heifer}
\alias{preg.heifer}
\title{Nutrient Requirements for Bred Heifers Calving as 2 Yr Old.}
\usage{
preg.heifer(breed.code = 1, mbw.lbs = 1200, breeding.pct.bw = 60,
  calving.pct.bw = 85)
}
\arguments{
\item{breed.code}{Default is 1 = Angus.  Breed code must match a code listed
in the table 19_1 data file.}

\item{mbw.lbs}{Default is 1200.  This value represents the mature, unshrunk bw
(lbs) for cows > 3 yrs old.  For first calf heifers, calving at 2 yr is assumed.
mbw.lbs is the estimated mature weight for first calf heifers, and equations are
based on #' a shrunk bw of 0.85(mbw) at first calving and 0.9(mbw) at second
calving.}

\item{breeding.pct.bw}{Default is 60. This value represents the percentage of
mature body weight at heifer breeding.}

\item{calving.pct.bw}{Default is 85. This value represents the target percentage
of mature body weight obtained by first calving.  This weight does not reflect the
additional weight of the conceptus.}
}
\value{
Two tables are printed including nutrient requirement totals and nutrient
requirement as a proportion of dry matter intake.
}
\description{
\code{preg.heifer} Nutrient requirements for bred heifers from
conception to calving.
}
\examples{
preg.heifer()
preg.heifer(breed.code=4,breeding.pct.bw=65)

}
