% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table191print.R
\name{Table19_1.print}
\alias{Table19_1.print}
\title{Print Table 19_1 Table Content}
\usage{
Table19_1.print()
}
\value{
A data frame with the content of Table 19_1 for breed.code reference.
}
\description{
Provides access to the internal data content of Table 19_1
}
