% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table19_1.R
\docType{data}
\name{table19_1}
\alias{table19_1}
\title{Table 19_1 NRC Breed Adjustments}
\format{An object of class \code{data.frame} with 24 rows and 9 columns.}
\usage{
table19_1
}
\description{
A dataset containing the adjustments to nutrient requirements for various breeds
}
\details{
\itemize{
   \item code. breed code
   \item breed.
   \item be. breed adjustment for NEm
   \item l. L.factor adjustment for NEm
   \item cbw. Average calf birth weight (kg)
   \item pkyd. Average Peak milk yield (kg)
   \item milk.fat. Milk fat percentage
   \item milk.prot. Milk protein percentage
   \item milk.solid. Milk solids percentage
   ...
   }
   @format A data frame with 24 rows and 9 variables
}
\keyword{datasets}
